app/src/main/java/com/example/assignment/data/MoviesPagingSource.kt
package com.example.assignment.data

import androidx.paging.PagingSource
import androidx.paging.PagingState
import com.example.assignment.api.MovieReviewApi
import retrofit2.HttpException
import java.io.IOException

private const val STARTING_PAGE_INDEX = 1

class MoviesPagingSource(
    private val movieReviewApi: MovieReviewApi
) : PagingSource<Int, Result>() {
    override suspend fun load(params: LoadParams<Int>): LoadResult<Int, Result> {
        val position = params.key ?: STARTING_PAGE_INDEX

        return try {
            val response = movieReviewApi.getMovies()
            val size = response.results.size
            val photos = response.results

            LoadResult.Page(
                data = photos,
                prevKey = if (position == STARTING_PAGE_INDEX) null else position - 1,
                nextKey = if (photos.size == size) null else position + 1
//
            )

        } catch (exception: IOException) {
            LoadResult.Error(exception)
        } catch (exception: HttpException) {
            LoadResult.Error(exception)
        }
    }

    override fun getRefreshKey(state: PagingState<Int, Result>): Int? {
        return state.anchorPosition
    }
}